git-svn-id: https://svn.janelia.org/eddylab/eddys/easel/branches/infernal/1.1@938 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ esl_sq.c
@@ -1421,7 +1421,7 @@ esl_sq_ReverseComplement(ESL_SQ *sq)
     }
 #endif /*eslAUGMENT_ALPHABET*/
 
-  ESL_SWAP(sq->start, sq->end, int);
+  ESL_SWAP(sq->start, sq->end, int64_t);
   /* revcomp invalidates any secondary structure annotation */
   if (sq->ss != NULL) { free(sq->ss); sq->ss = NULL; }
   return status;
